Switching to FP16

+++ keras_retinanet/trt_conversion/to_trt.py
@@ -48,7 +48,7 @@ with tf.Session() as sess:
         outputs=output_names,
         max_batch_size=1,
         max_workspace_size_bytes=1 << 25,
-        precision_mode='INT8',
+        precision_mode='FP16',
         is_dynamic_op=False,
         minimum_segment_size=5
     )
